add command-line executable option to replace_members_with_ids
